Rename etsy to etsy2
The package is published on pypi as `etsy2`. This commit updates
the package so that it's available to import under the same name.
